requirements.txt
requests
bs4
python-dotenv
openai
trafilatura
tldextract